iteration fibbonachi

+++ AL01A/src/ee/ttu/algoritmid/fibonacci/AL01A.java
@@ -10,13 +10,13 @@ public class AL01A {
      * @param n The number of the sequence to compute.
      * @return The n-th number in Fibonacci series.
      */
-    public String iterativeF(final Integer n) {
-        Integer[] f = new Integer[n + 1];
+    public String iterativeF(final long n) {
+        long[] f = new long[(int) (n + 1)];
         f[0] = 1;
         f[1] = 1;
         for (int i = 2; i <= n; i++) {
             f[i] = f[i - 1] + f[i - 2];
         }
-        return String.valueOf(f[n]);
+        return String.valueOf(f[(int) n]);
     }
 }
